+++ src/Data/Tuple/Append/Example.hs
@@ -1,12 +1,12 @@
-{-# LANGUAGE MagicHash, UnboxedTuples, TemplateHaskell #-}
+{-# LANGUAGE MagicHash, TemplateHaskell, UnboxedTuples #-}
 {-# OPTIONS_GHC -fobject-code #-}
 
 module Data.Tuple.Append.Example where
 
 import Data.Tuple.Append.TemplateHaskell(makeUnboxedTupleAppendFun)
 
-import GHC.Exts(Int#, Float#)
+import GHC.Exts(Float#, Int#)
 
 import Language.Haskell.TH.Syntax(Type(ConT), mkName)
 
-makeUnboxedTupleAppendFun (mkName "foo") [ ConT ''Int#, ConT ''Float# ] [ ConT ''Float#, ConT ''Int# ]
+makeUnboxedTupleAppendFun (mkName "append_if_fi") [ ConT ''Int#, ConT ''Float# ] [ ConT ''Float#, ConT ''Int# ]