+++ CHANGELOG.md
@@ -4,4 +4,4 @@ For a full list of changes, see the history on [*GitHub*](https://github.com/hap
 
 ## Version 0.1.0.0
 
-
+A first intial version that allows constructing arbitrary prepend and append functions and typeclass instances. Instances for `AddL` and `AddR` are defined to construct tuples with a length up to 62, and it is possible to construct any tuple by appending tuples for a size up to 19.